move 'paginate()' from BlogPage component to PostsContainer
